+++ admin/src/Scheduling/index.js
@@ -1,79 +1,27 @@
 import React from 'react';
-import stylish from '@dmamills/stylish';
 import cn from 'classnames';
 
 import { getSchedules } from '../api';
+import Table from './Table'
 
-import { p1, flex, spaceBetween, flexCenter, width100, justifyEnd } from '../styles';
-
-const [ table, tableHeader, tableBody ] = stylish({
-  borderCollapse: 'collapse',
-}, {
-  backgroundColor: 'black',
-  color: 'white',
-  padding: '1rem',
-  ' > th': {
-    padding: '1rem',
-    textAlign: 'left',
-  },
-}, {
-  color: 'black',
-  ':nth-child(odd)': {
-    backgroundColor: 'lightgrey'
-  },
-  ':nth-child(even)': {
-    backgroundColor: '#eee'
-  },
-  ' > td': {
-    padding: '1rem'
-  }
-});
-
-const Row = ({ schedule }) => {
-  return (<tr className={cn(tableBody, p1)}>
-    <td>{schedule.id}</td>
-    <td>{schedule.name}</td>
-    <td>{schedule.description}</td>
-    <td>{schedule.start_time}</td>
-    <td>{schedule.length} hour{schedule.length > 1 ? 's' : ''}</td>
-    <td className={cn(flex, justifyEnd)}>
-      <button>Edit</button>
-      <button>Delete</button>
-    </td>
-  </tr>)
-}
-
-const Table = ({ schedules }) => {
-  return (<table className={cn(width100, table)}>
-    <thead className={tableHeader}>
-      <tr>
-        <th>Id</th>
-        <th>Name</th>
-        <th>Description</th>
-        <th>Start Time</th>
-        <th>Length</th>
-        <th></th>
-      </tr>
-    </thead>
-    <tbody>
-      {schedules.map(s => <Row key={s.id} schedule={s} />)}
-    </tbody>
-  </table>)
-}
-
+import { p1, flex, spaceBetween, flexCenter } from '../styles';
 
 class Scheduling extends React.Component {
   state = {
     schedules: []
   }
 
-  componentDidMount() {
+  fetchSchedules = () => {
     getSchedules()
     .then(schedules => {
       this.setState({
         schedules
-      })
-    })
+      });
+    });
+  }
+
+  componentDidMount() {
+   this.fetchSchedules();
   }
 
   render() {
@@ -84,13 +32,16 @@ class Scheduling extends React.Component {
           <h2>Scheduling</h2>
           <div className={cn(flex, flexCenter)}>
             <button>Create New Schedule</button>
+            <button onClick={this.fetchSchedules}>Refresh</button>
           </div>
         </div>
         <div>
-          <Table schedules={schedules}/>
+          <Table
+            schedules={schedules}
+          />
         </div>
       </div>
-    )
+    );
   }
 }
